+++ CppWinForm2/Matrix.h
@@ -2,20 +2,20 @@ class MATRIX
 {
 private:
 	int row, col;
-	float **p;
+	double **p;
 public:		
 	MATRIX(int, int);
-	void Set(int, int, float) const;
-	float Get(int, int) const;
+	void Set(int, int, double) const;
+	double Get(int, int) const;
 	int getRow() const;
 	int getColumn() const;
 	MATRIX rowSwap(int row1, int row2);
 	MATRIX pivotBelirle(int * pivot);
 	MATRIX pivotKullan(int pivot);
 
-	MATRIX MATRIX::SatirBol(int satir, float islem);
+	MATRIX MATRIX::SatirBol(int satir, double islem);
 	
-	MATRIX SatirIslem(int satir, int satir2, float islem); 
+	MATRIX SatirIslem(int satir, int satir2, double islem); 
 	MATRIX operator / (int n);
 	bool operator ==(const MATRIX& other);
 	bool operator!=(const MATRIX& other);
